# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger api docs",
        "title": "GameObjectBackend",
        "contact": {
            "name": "WayneChu",
            "url": "https://github.com/Onetail"
        },
        "license": {
            "name": "Nginx",
            "url": "https://www.nginx.com/"
        },
        "version": "1.0.0"
    },
    "host": "localhost:7011",
    "paths": {
        "/api/v1/accounts/signin": {
            "post": {
                "description": "回傳 user token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "get user token",
                "parameters": [
                    {
                        "description": "參數",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PostSigninBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.UserSignInResponseObject"
                        }
                    },
                    "403": {
                        "description": "err.Error()",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/users/": {
            "get": {
                "description": "取得 user 列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user list",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.UserListResponseObject"
                        }
                    },
                    "403": {
                        "description": "err.Error()",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "新增 user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "create user",
                "parameters": [
                    {
                        "description": "參數",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateUserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        },
        "/api/v1/users/:userId": {
            "get": {
                "description": "取得單一 user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user by userId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.UserResponseObject"
                        }
                    },
                    "404": {
                        "description": "record not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "刪除 user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/model.RowsAffectedModel"
                        }
                    }
                }
            },
            "patch": {
                "description": "更新 user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "update user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "參數",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateUserBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/model.RowsAffectedModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateUserBody": {
            "type": "object",
            "required": [
                "birthday",
                "email",
                "gender",
                "nickname",
                "password",
                "phoneCountryCode",
                "phoneNumber"
            ],
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phoneCountryCode": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                }
            }
        },
        "model.PostSigninBody": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.RowsAffectedModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                }
            }
        },
        "model.UpdateUserBody": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phoneCountryCode": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.UserListResponseObject": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.UserResponseObject": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.UserSignInResponseObject": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                }
            }
        }
    }
}